move config to util
